// src/content/types.ts
export type IconName =
  | "globe"
  | "smartphone"
  | "blocks"
  | "bot"
  | "audio"
  | "message"
  | "workflow"
  | "cloud"
  | "graduation"
  | "hotel"
  | "factory"
  | "building"
  | "heart"
  | "briefcase"
  | "compass"
  | "cpu"
  | "layers"
  | "shield"
  | "handshake"
  | "link"
  | "database";